Fixed the missile movement Fixed incorrect landing positions Added missile entity Fixed conversation duping

+++ src/main/java/me/kaiyan/missilewarfare/Items/MissileClass.java
@@ -1,13 +1,13 @@
 package me.kaiyan.missilewarfare.Items;
 
 public class MissileClass {
-    public int speed;
+    public float speed;
     public int range;
     public double power;
     public int accuracy;
     public int type;
 
-    public MissileClass(int speed,int range, double power, int accuracy, int type){
+    public MissileClass(float speed,int range, double power, int accuracy, int type){
        this.speed = speed;
        this.range = range;
        this.power = power;
